Add documentation, rename isomorphic to hydration & remove document template
