+++ app/layouts/default_layout.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Helmet } from 'react-helmet'
-import favicon from '../images/favicon.png'
+import favicon from './favicon.png'
 
 export default function DefaultLayout ({ meta, children }) {
   return (
